+++ src/test/java/lotto/LottoPurchaseAmountValidatorTest.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.exception.PurchaseAmountExceptionMessage;
+import lotto.validator.AmountValidator;
+import lotto.validator.LottoPurchaseAmountValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
